@@ -1,16 +1,20 @@
 package nodomain.sakiika.ranamod.datagen.loot;
 
+import net.minecraft.core.Holder;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.flag.FeatureFlags;
 import net.minecraft.data.loot.EntityLootSubProvider;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.level.storage.loot.entries.LootItem;
-import net.minecraft.world.level.storage.loot.functions.LootingEnchantFunction;
+import net.minecraft.world.level.storage.loot.functions.EnchantedCountIncreaseFunction;
 import net.minecraft.world.level.storage.loot.functions.SetItemCountFunction;
+import net.minecraft.world.level.storage.loot.providers.number.BinomialDistributionGenerator;
 import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
+import net.minecraft.world.level.storage.loot.providers.number.NumberProviders;
 import net.minecraft.world.level.storage.loot.providers.number.UniformGenerator;
-import net.minecraftforge.registries.RegistryObject;
+
 import nodomain.sakiika.ranamod.entity.ModEntities;
 
 import java.util.stream.Stream;
@@ -18,8 +22,10 @@ import java.util.stream.Stream;
 import static net.minecraft.world.level.storage.loot.LootPool.lootPool;
 
 public class ModEntityLootTableProvider extends EntityLootSubProvider {
-    public ModEntityLootTableProvider() {
-        super(FeatureFlags.REGISTRY.allFlags());
+    private HolderLookup.Provider provider;
+    public ModEntityLootTableProvider(HolderLookup.Provider provider) {
+        super(FeatureFlags.REGISTRY.allFlags(), provider);
+        this.provider = provider;
     }
 
     @Override
@@ -32,7 +38,7 @@ public class ModEntityLootTableProvider extends EntityLootSubProvider {
                                 .apply(SetItemCountFunction.setCount(
                                         UniformGenerator.between(1.0F, 3.0F) //amount of 1-3
                                 ))
-                                .apply(LootingEnchantFunction.lootingMultiplier(
+                                .apply(EnchantedCountIncreaseFunction.lootingMultiplier(provider,
                                         UniformGenerator.between(0.0F, 1.0F) //increasing amount by looting enchant
                                 ))
                                 .setWeight(975) //97.5% chance to drop
@@ -42,7 +48,7 @@ public class ModEntityLootTableProvider extends EntityLootSubProvider {
                                 .apply(SetItemCountFunction.setCount(
                                         ConstantValue.exactly(1.0F)
                                 ))
-                                .apply(LootingEnchantFunction.lootingMultiplier(
+                                .apply(EnchantedCountIncreaseFunction.lootingMultiplier(provider,
                                         UniformGenerator.between(0.0F, 1.0F)
                                 ))
                                 .setWeight(20) //2% chance to drop
@@ -52,7 +58,7 @@ public class ModEntityLootTableProvider extends EntityLootSubProvider {
                                 .apply(SetItemCountFunction.setCount(
                                         ConstantValue.exactly(1.0F)
                                 ))
-                                .apply(LootingEnchantFunction.lootingMultiplier(
+                                .apply(EnchantedCountIncreaseFunction.lootingMultiplier(provider,
                                         UniformGenerator.between(0.0F, 1.0F)
                                 ))
                                 .setWeight(5) //0.5% chance to drop
@@ -62,6 +68,6 @@ public class ModEntityLootTableProvider extends EntityLootSubProvider {
 
     @Override
     protected Stream<EntityType<?>> getKnownEntityTypes() {
-        return ModEntities.ENTITY_TYPES.getEntries().stream().map(RegistryObject::get);
+        return ModEntities.ENTITY_TYPES.getEntries().stream().map(Holder::value);
     }
 }
