@@ -1,19 +1,20 @@
 package nodomain.sakiika.ranamod.entity;
 
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.neoforge.registries.DeferredRegister;
 import nodomain.sakiika.ranamod.RanaMod;
 import nodomain.sakiika.ranamod.entity.custom.RanaEntity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobCategory;
-import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.registries.DeferredRegister;
-import net.minecraftforge.registries.ForgeRegistries;
-import net.minecraftforge.registries.RegistryObject;
+
+import java.util.function.Supplier;
 
 public class ModEntities {
     public static final DeferredRegister<EntityType<?>> ENTITY_TYPES =
-            DeferredRegister.create(ForgeRegistries.ENTITY_TYPES, RanaMod.MOD_ID);
+            DeferredRegister.create(BuiltInRegistries.ENTITY_TYPE, RanaMod.MOD_ID);
 
-    public static final RegistryObject<EntityType<RanaEntity>> RANA =
+    public static final Supplier<EntityType<RanaEntity>> RANA =
             ENTITY_TYPES.register("rana", () -> EntityType.Builder.of(RanaEntity::new, MobCategory.CREATURE)
                     .sized(0.5f, 1.5f).build("rana")); // sized means boundary box
 
